Little work on the main window

+++ game_of_life/pre_pray/src/Food.h
@@ -17,7 +17,7 @@ namespace game_of_life
     std::shared_ptr<Food> ptr(){return shared_from_this();}
     virtual uint8_t health(void)const;
     private:
-      float _amount;
+      unsigned int _amount;
 };
 }
 }
